# tests/fixtures/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Article Analysis API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/runs/{run_id}": {
      "get": {
        "tags": [
          "runs"
        ],
        "summary": "Get Run",
        "operationId": "get_run_api_v1_runs__run_id__get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunV1"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/runs/{run_id}/cancel": {
      "post": {
        "tags": [
          "runs"
        ],
        "summary": "Cancel",
        "operationId": "cancel_api_v1_runs__run_id__cancel_post",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Cancel Api V1 Runs  Run Id  Cancel Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/runs/{run_id}/evidence/{evidence_id}": {
      "get": {
        "tags": [
          "runs"
        ],
        "summary": "Evidence",
        "operationId": "evidence_api_v1_runs__run_id__evidence__evidence_id__get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Run Id"
            }
          },
          {
            "name": "evidence_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Evidence Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvidenceV1"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/runs/{run_id}/events": {
      "get": {
        "tags": [
          "runs"
        ],
        "summary": "Events",
        "operationId": "events_api_v1_runs__run_id__events_get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Run Id"
            }
          },
          {
            "name": "last-event-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Last-Event-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "text/event-stream": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AnswerPresentationV1": {
        "properties": {
          "schema_version": {
            "type": "integer",
            "const": 1,
            "title": "Schema Version",
            "default": 1
          },
          "presentation_id": {
            "type": "string",
            "pattern": "^[0-9a-f]{64}$",
            "title": "Presentation Id"
          },
          "renderer_version": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "title": "Renderer Version"
          },
          "text": {
            "type": "string",
            "maxLength": 65536,
            "title": "Text"
          },
          "text_sha256": {
            "type": "string",
            "pattern": "^[0-9a-f]{64}$",
            "title": "Text Sha256"
          },
          "chunk_count": {
            "type": "integer",
            "maximum": 128.0,
            "minimum": 1.0,
            "title": "Chunk Count"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "presentation_id",
          "renderer_version",
          "text",
          "text_sha256",
          "chunk_count"
        ],
        "title": "AnswerPresentationV1"
      },
      "AnswerV1": {
        "properties": {
          "schema_version": {
            "type": "integer",
            "const": 1,
            "title": "Schema Version",
            "default": 1
          },
          "summary": {
            "items": {
              "$ref": "#/components/schemas/ClaimV1"
            },
            "type": "array",
            "title": "Summary"
          },
          "matches": {
            "items": {
              "$ref": "#/components/schemas/MatchV1"
            },
            "type": "array",
            "title": "Matches"
          },
          "differences": {
            "items": {
              "$ref": "#/components/schemas/DifferenceV1"
            },
            "type": "array",
            "title": "Differences"
          },
          "limitations": {
            "items": {
              "$ref": "#/components/schemas/LimitationV1"
            },
            "type": "array",
            "title": "Limitations"
          },
          "followup_suggestions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Followup Suggestions"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "summary",
          "matches",
          "differences",
          "limitations",
          "followup_suggestions"
        ],
        "title": "AnswerV1"
      },
      "ChannelCoverageV1": {
        "properties": {
          "channel": {
            "type": "string",
            "title": "Channel"
          },
          "status": {
            "type": "string",
            "enum": [
              "ok",
              "empty",
              "unavailable",
              "disabled"
            ],
            "title": "Status"
          },
          "reason_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason Code"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "channel",
          "status"
        ],
        "title": "ChannelCoverageV1"
      },
      "ClaimV1": {
        "properties": {
          "text": {
            "type": "string",
            "maxLength": 4000,
            "minLength": 1,
            "title": "Text"
          },
          "evidence_ids": {
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "type": "array",
            "maxItems": 64,
            "minItems": 1,
            "title": "Evidence Ids"
          },
          "quotes": {
            "items": {
              "$ref": "#/components/schemas/QuoteV1"
            },
            "type": "array",
            "maxItems": 64,
            "minItems": 1,
            "title": "Quotes"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "text",
          "evidence_ids",
          "quotes"
        ],
        "title": "ClaimV1"
      },
...
      "CoverageV1": {
        "properties": {
          "sources": {
            "items": {
              "$ref": "#/components/schemas/SourceCoverageV1"
            },
            "type": "array",
            "title": "Sources"
          },
          "channels": {
            "items": {
              "$ref": "#/components/schemas/ChannelCoverageV1"
            },
            "type": "array",
            "title": "Channels"
          },
          "partial": {
            "type": "boolean",
            "title": "Partial"
          },
          "historical": {
            "type": "boolean",
            "title": "Historical"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "sources",
          "channels",
          "partial",
          "historical"
        ],
        "title": "CoverageV1"
      },
      "DifferenceV1": {
        "properties": {
          "feature_id": {
            "type": "string",
            "format": "uuid",
            "title": "Feature Id"
          },
          "claims": {
            "items": {
              "$ref": "#/components/schemas/ClaimV1"
            },
            "type": "array",
            "title": "Claims"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "feature_id",
          "claims"
        ],
        "title": "DifferenceV1"
      },
      "EvidenceV1": {
        "properties": {
          "evidence_id": {
            "type": "string",
            "format": "uuid",
            "title": "Evidence Id"
          },
          "document_id": {
            "type": "string",
            "format": "uuid",
            "title": "Document Id"
          },
          "revision_id": {
            "type": "string",
            "format": "uuid",
            "title": "Revision Id"
          },
          "chunk_id": {
            "type": "string",
            "format": "uuid",
            "title": "Chunk Id"
          },
          "section": {
            "type": "string",
            "title": "Section"
          },
          "span_start": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Span Start"
          },
          "span_end": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Span End"
          },
          "quoted_span": {
            "type": "string",
            "title": "Quoted Span"
          },
          "source_url": {
            "type": "string",
            "title": "Source Url"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "evidence_id",
          "document_id",
          "revision_id",
          "chunk_id",
          "section",
          "span_start",
          "span_end",
          "quoted_span",
          "source_url"
        ],
        "title": "EvidenceV1"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "LimitationV1": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "LimitationV1"
      },
      "MatchV1": {
        "properties": {
          "feature_id": {
            "type": "string",
            "format": "uuid",
            "title": "Feature Id"
          },
          "document_id": {
            "type": "string",
            "format": "uuid",
            "title": "Document Id"
          },
          "claims": {
            "items": {
              "$ref": "#/components/schemas/ClaimV1"
            },
            "type": "array",
            "title": "Claims"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "feature_id",
          "document_id",
          "claims"
        ],
        "title": "MatchV1"
      },
      "ProgressCountsV1": {
        "additionalProperties": {
          "type": "integer"
        },
        "type": "object"
      },
      "ProgressV1": {
        "properties": {
          "schema_version": {
            "type": "integer",
            "const": 1,
            "title": "Schema Version",
            "default": 1
          },
          "attempt": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Attempt"
          },
          "stage": {
            "type": "string",
            "title": "Stage"
          },
          "phase": {
            "type": "string",
            "enum": [
              "started",
              "completed",
              "waiting"
            ],
            "title": "Phase"
          },
          "stage_started_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stage Started At"
          },
          "counts": {
            "$ref": "#/components/schemas/ProgressCountsV1"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "attempt",
          "stage",
          "phase",
          "counts"
        ],
        "title": "ProgressV1"
      },
      "PublicAnalysisV1": {
        "properties": {
          "schema_version": {
            "type": "integer",
            "const": 1,
            "title": "Schema Version",
            "default": 1
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/ClaimV1"
            },
            "type": "array",
            "maxItems": 12,
            "title": "Items"
          },
          "limitations": {
            "items": {
              "$ref": "#/components/schemas/LimitationV1"
            },
            "type": "array",
            "maxItems": 16,
            "title": "Limitations"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "items",
          "limitations"
        ],
        "title": "PublicAnalysisV1"
      },
      "QuoteV1": {
        "properties": {
          "evidence_id": {
            "type": "string",
            "format": "uuid",
            "title": "Evidence Id"
          },
          "start": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Start"
          },
          "end": {
            "type": "integer",
            "minimum": 0.0,
            "title": "End"
          },
          "text": {
            "type": "string",
            "title": "Text"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "evidence_id",
          "start",
          "end",
          "text"
        ],
        "title": "QuoteV1"
      },
      "RunV1": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "running",
              "completed",
              "failed",
              "cancelled"
            ],
            "title": "Status"
          },
          "stage": {
            "type": "string",
            "title": "Stage"
          },
          "progress": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ProgressV1"
              },
              {
                "additionalProperties": true,
                "type": "object"
              }
            ],
            "title": "Progress",
            "default": {}
          },
          "public_analysis": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PublicAnalysisV1"
              },
              {
                "type": "null"
              }
            ]
          },
          "answer_presentation": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AnswerPresentationV1"
              },
              {
                "type": "null"
              }
            ]
          },
          "idea_version_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Idea Version Id"
          },
          "source_run_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Run Id"
          },
          "outcome": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "analysis",
                  "safe_fallback",
                  "no_evidence",
                  "clarification"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Outcome"
          },
          "answer": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AnswerV1"
              },
              {
                "type": "null"
              }
            ]
          },
          "sources": {
            "items": {
              "$ref": "#/components/schemas/SourceV1"
            },
            "type": "array",
            "title": "Sources"
          },
          "coverage": {
            "$ref": "#/components/schemas/CoverageV1"
          },
          "graph_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Graph Url"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "completed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "error_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Code"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "status",
          "stage",
          "sources",
          "coverage",
          "created_at"
        ],
        "title": "RunV1"
      },
...
      "SourceCoverageV1": {
        "properties": {
          "source": {
            "type": "string",
            "title": "Source"
          },
          "status": {
            "type": "string",
            "enum": [
              "ok",
              "empty",
              "unavailable",
              "not_configured",
              "not_requested"
            ],
            "title": "Status"
          },
          "reason_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason Code"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "source",
          "status"
        ],
        "title": "SourceCoverageV1"
      },
      "SourceV1": {
        "properties": {
          "document_id": {
            "type": "string",
            "format": "uuid",
            "title": "Document Id"
          },
          "revision_id": {
            "type": "string",
            "format": "uuid",
            "title": "Revision Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "kind": {
            "type": "string",
            "title": "Kind"
          },
          "publication_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Publication Date"
          },
          "url": {
            "type": "string",
            "title": "Url"
          },
          "evidence_ids": {
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "type": "array",
            "title": "Evidence Ids"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "document_id",
          "revision_id",
          "title",
          "kind",
          "url",
          "evidence_ids"
        ],
        "title": "SourceV1"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}